qwen_tts_service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
torch
torchaudio
soundfile
numpy
scipy
librosa
denoiser
julius
novasr @ git+https://github.com/ysharma3501/NovaSR.git
einops
soxr
qwen-omni-utils
nagisa
soynlp
pytz
redis
pyannote.audio
